requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-login>=0.6.3
flask-bcrypt>=1.0.1
flask-wtf>=1.2.1
flask-jwt-extended>=4.5.3
wtforms>=3.1.0
openai>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
transformers>=4.35.0
torch>=2.6.0
schedule>=1.2.0
email-validator>=2.1.0
textblob>=0.17.1
langdetect>=1.0.9
twilio>=8.10.0
telebot>=0.0.5
pyTelegramBotAPI>=4.14.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
flask-sqlalchemy>=3.1.1
flask-migrate>=4.0.5
speechrecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.11
facebook-sdk>=3.1.0
openpyxl>=3.1.2
pandas>=2.1.1
numpy>=1.25.2
scipy>=1.11.3
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.16.1
pillow>=10.0.1
qrcode>=7.4.2
hubspot-api-client>=7.0.0
mailchimp3>=3.0.21